Abstracted behaviour from three labware creators into a new partial stamped plate superclass.
