+++ app/models/labware_creators/normalised_binned_plate.rb
@@ -13,47 +13,12 @@ module LabwareCreators
   # to further normalise the samples.
   # The library preparation Submission may only be for some of the wells on the plate. We use a well
   # filter on library type to select only those wells that have a request for the required library type.
-  class NormalisedBinnedPlate < StampedPlate
-    include LabwareCreators::RequireWellsWithConcentrations
-    include LabwareCreators::GenerateQCResults
-
-    validate :wells_with_aliquots_have_concentrations?
-    validate :transfer_hash_present?
-    validate :number_of_transfers_matches_number_of_filtered_wells?
-
+  class NormalisedBinnedPlate < PartialStampedPlate
     def dilutions_calculator
       @dilutions_calculator ||= Utility::NormalisedBinningCalculator.new(dilutions_config)
     end
 
     private
-
-    # use well filter variant that allows for partial submissions of a plate
-    def well_filter
-      @well_filter ||= WellFilterAllowingPartials.new(creator: self, request_state: 'pending')
-    end
-
-    def filtered_wells
-      well_filter.filtered.each_with_object([]) do |well_filter_details, wells|
-        wells << well_filter_details[0]
-      end
-    end
-
-    # Validation to check we have identified wells to transfer.
-    # Plate must contain at least one well with a request for library preparation, in a state of pending.
-    def transfer_hash_present?
-      return if transfer_hash.present?
-
-      msg = 'No wells have pending library preparation requests with the expected library type (check the submission).'
-      errors.add(:parent, msg)
-    end
-
-    # Validation to check number of filtered wells matches to final transfers hash produced
-    def number_of_transfers_matches_number_of_filtered_wells?
-      return if transfer_hash.length == filtered_wells.length
-
-      msg = 'Number of filtered wells does not match number of well transfers'
-      errors.add(:parent, msg)
-    end
 
     def request_hash(source_well, child_plate, additional_parameters)
       return unless transfer_hash.key?(source_well.location)
@@ -65,10 +30,6 @@ module LabwareCreators
         end&.uuid,
         'volume' => transfer_hash[source_well.location]['volume'].to_s
       }.merge(additional_parameters)
-    end
-
-    def transfer_hash
-      @transfer_hash ||= dilutions_calculator.compute_well_transfers(parent, filtered_wells)
     end
   end
 end
